feat: seperating test cases for deepresearch and ai_scrape.
